refactor: 시작 파일과 메인 페이지 분리
